Move `ov_phi3_vision.py` from engine folder to mutlimodal_models/openvino
