+++ src/embeddedllm/multimodal_models/openvino/ov_phi3_vision.py
